adding a  python module and adding the possibility of chosing the input  datasets
